+++ platform-api/src/main/java/com/flow/platform/api/dao/user/UserFlowDao.java
@@ -20,7 +20,6 @@ import com.flow.platform.api.domain.user.UserFlowKey;
 import com.flow.platform.core.dao.BaseDao;
 import com.flow.platform.core.domain.Page;
 import com.flow.platform.core.domain.Pageable;
-import java.awt.PageAttributes;
 import java.util.List;
 
 /**
