+++ platform-api/src/test/java/com/flow/platform/api/test/dao/FlowDaoTest.java
@@ -21,7 +21,6 @@ import com.flow.platform.api.domain.node.Node;
 import com.flow.platform.api.test.TestBase;
 import com.flow.platform.core.domain.Page;
 import com.flow.platform.core.domain.Pageable;
-import com.flow.platform.core.domain.PageableImpl;
 import com.google.common.collect.Sets;
 import java.time.ZonedDateTime;
 import java.util.List;
@@ -59,7 +58,7 @@ public class FlowDaoTest extends TestBase {
     @Test
     public void should_list_page_success(){
 
-        Pageable pageable = new PageableImpl(0, 10);
+        Pageable pageable = new Pageable(0, 10);
         Page<Node> page = flowDao.list(pageable);
         Assert.assertNotNull(page);
         Assert.assertEquals(1, page.getContent().size());
@@ -105,7 +104,7 @@ public class FlowDaoTest extends TestBase {
 
     @Test
     public void should_list_path_page_by_created_by() {
-        Pageable pageable = new PageableImpl(0, 10);
+        Pageable pageable = new Pageable(0, 10);
         Page<String> page = flowDao.pathList(Sets.newHashSet("admin@flow.ci"), pageable);
         Assert.assertEquals(1, page.getContent().size());
         Assert.assertEquals(flow.getPath(), page.getContent().get(0));
