--- platform-core/src/main/java/com/flow/platform/core/domain/PageableImpl.java
@@ -1,60 +1,0 @@
-/*
- * Copyright 2017 flow.ci
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.flow.platform.core.domain;
-
-/**
- * @author gyfirim
- */
-public class PageableImpl implements Pageable {
-
-    private int number;
-
-    private int size;
-
-    public PageableImpl(int number, int size) {
-        this.number = number;
-        this.size = size;
-    }
-
-    public PageableImpl (){
-
-    }
-
-    @Override
-    public int getPageNumber() {
-        return number;
-    }
-
-    @Override
-    public int getPageSize() {
-        return size;
-    }
-
-    @Override
-    public int getOffset() {
-        return (number-1) * size;
-    }
-
-    public void setSize(int size) {
-        this.size = size;
-    }
-
-    public void setNumber(int number) {
-        this.number = number;
-    }
-
-}
